+++ core-parent/admin-core-biz/src/main/java/com/kalix/admin/core/biz/WorkGroupBeanServiceImpl.java
@@ -18,6 +18,7 @@ import java.util.List;
 
 /**
  * 工作组管理服务实现
+ *
  * @author majian <br/>
  *         date:2015-7-27
  * @version 1.0.0
@@ -62,11 +63,11 @@ public class WorkGroupBeanServiceImpl extends ShiroGenericBizServiceImpl<IWorkGr
     @Override
     public boolean isUpdate(WorkGroupBean entity, JsonStatus status) {
         Assert.notNull(entity, "实体不能为空.");
-        WorkGroupBean bean=(WorkGroupBean)entity;
+        WorkGroupBean bean = (WorkGroupBean) entity;
         List<WorkGroupBean> beans = dao.find("select ob from WorkGroupBean ob where ob.name = ?1", bean.getName());
-        if(beans!=null&&beans.size()>0){
-            WorkGroupBean _workGroup=beans.get(0);
-            if(_workGroup.getId()!=entity.getId()) {
+        if (beans != null && beans.size() > 0) {
+            WorkGroupBean _workGroup = beans.get(0);
+            if (_workGroup.getId() != entity.getId()) {
                 status.setFailure(true);
                 status.setMsg(FUNCTION_NAME + "已经存在！");
                 return false;
@@ -78,9 +79,9 @@ public class WorkGroupBeanServiceImpl extends ShiroGenericBizServiceImpl<IWorkGr
     @Override
     public boolean isSave(WorkGroupBean entity, JsonStatus status) {
         Assert.notNull(entity, "实体不能为空.");
-        WorkGroupBean bean=(WorkGroupBean)entity;
+        WorkGroupBean bean = (WorkGroupBean) entity;
         List<WorkGroupBean> beans = dao.find("select ob from WorkGroupBean ob where ob.name = ?1", bean.getName());
-        if(beans!=null&&beans.size()>0){
+        if (beans != null && beans.size() > 0) {
             status.setSuccess(false);
             status.setMsg(FUNCTION_NAME + "已经存在！");
             return false;
@@ -89,30 +90,31 @@ public class WorkGroupBeanServiceImpl extends ShiroGenericBizServiceImpl<IWorkGr
     }
 
     @Override
-    public JsonData getAllWorkGroup(int page,int limit) {
+    public JsonData getAllWorkGroup(int page, int limit) {
         return dao.getAll(page, limit);
     }
 
     @Override
-    public List getUsersByWorkGroupId(long id) {
-        List userIds=new ArrayList<String>();
-        List<WorkGroupUserBean> workGroupUserBeans=workGroupUserBeanDao.find("select ob from WorkGroupUserBean ob where ob.groupId = ?1", id);
-        if(workGroupUserBeans!=null&&!workGroupUserBeans.isEmpty()){
-            for(WorkGroupUserBean workGroupUserBean:workGroupUserBeans){
-                if(workGroupUserBean!=null&&workGroupUserBean.getUserId()!=0){
+    public List getUserIdsByWorkGroupId(long id) {
+        List userIds = new ArrayList<String>();
+        List<WorkGroupUserBean> workGroupUserBeans = workGroupUserBeanDao.find("select ob from WorkGroupUserBean ob where ob.groupId = ?1", id);
+        if (workGroupUserBeans != null && !workGroupUserBeans.isEmpty()) {
+            for (WorkGroupUserBean workGroupUserBean : workGroupUserBeans) {
+                if (workGroupUserBean != null && workGroupUserBean.getUserId() != 0) {
                     userIds.add(workGroupUserBean.getUserId());
                 }
             }
         }
         return userIds;
     }
+
     @Override
-    public List getRolesByWorkGroupId(long id) {
-        List roleIds=new ArrayList<>();
-        List<WorkGroupRoleBean> workGroupRoleBeans=workGroupRoleBeanDao.find("select ob from WorkGroupRoleBean ob where ob.groupId = ?1", id);
-        if(workGroupRoleBeans!=null&&!workGroupRoleBeans.isEmpty()){
-            for(WorkGroupRoleBean workGroupRoleBean:workGroupRoleBeans){
-                if(workGroupRoleBean!=null&&workGroupRoleBean.getRoleId()!=0){
+    public List getRoleIdsByWorkGroupId(long id) {
+        List roleIds = new ArrayList<>();
+        List<WorkGroupRoleBean> workGroupRoleBeans = workGroupRoleBeanDao.find("select ob from WorkGroupRoleBean ob where ob.groupId = ?1", id);
+        if (workGroupRoleBeans != null && !workGroupRoleBeans.isEmpty()) {
+            for (WorkGroupRoleBean workGroupRoleBean : workGroupRoleBeans) {
+                if (workGroupRoleBean != null && workGroupRoleBean.getRoleId() != 0) {
                     roleIds.add(workGroupRoleBean.getRoleId());
                 }
             }
@@ -121,13 +123,21 @@ public class WorkGroupBeanServiceImpl extends ShiroGenericBizServiceImpl<IWorkGr
     }
 
     @Override
-    public JsonStatus saveWorkGroupUsers(long workGroupId, String userId) {
-        JsonStatus jsonStatus=new JsonStatus();
-        try {
-            workGroupUserBeanDao.deleteByWorkGroupId(workGroupId);
-            String userName=getShiroService().getCurrentUserName();
-            if (StringUtils.isNotEmpty(userId)) {
-                if (userId.indexOf(",") != -1) {
+    public JsonStatus saveWorkGroupUsers(List ids) {
+        JsonStatus jsonStatus = new JsonStatus();
+
+        if (ids == null || ids.size() != 2) {
+            jsonStatus.setFailure(true);
+            jsonStatus.setMsg("保存失败!");
+            return jsonStatus;
+        } else {
+            try {
+                long workGroupId = Long.valueOf(ids.get(0).toString());
+                String userId = ids.get(1).toString();
+
+                workGroupUserBeanDao.deleteByWorkGroupId(workGroupId);
+                String userName = getShiroService().getCurrentUserName();
+                if (StringUtils.isNotEmpty(userId)) {
                     String[] userIds = userId.split(",");
                     for (String _userId : userIds) {
                         if (StringUtils.isNotEmpty(_userId.trim())) {
@@ -139,36 +149,37 @@ public class WorkGroupBeanServiceImpl extends ShiroGenericBizServiceImpl<IWorkGr
                             workGroupUserBeanDao.save(workGroupUserBean);
                         }
                     }
-                }else{
-                    if (StringUtils.isNotEmpty(userId.trim())) {
-                        WorkGroupUserBean workGroupUserBean = new WorkGroupUserBean();
-                        workGroupUserBean.setCreateBy(userName);
-                        workGroupUserBean.setUpdateBy(userName);
-                        workGroupUserBean.setGroupId(workGroupId);
-                        workGroupUserBean.setUserId(Long.parseLong(userId));
-                        workGroupUserBeanDao.save(workGroupUserBean);
-                    }
                 }
+            } catch (Exception e) {
+                e.printStackTrace();
+                jsonStatus.setFailure(true);
+                jsonStatus.setMsg("保存失败!");
+                return jsonStatus;
             }
-        }catch(Exception e){
-            e.printStackTrace();
-            jsonStatus.setFailure(true);
-            jsonStatus.setMsg("保存失败!");
-            return jsonStatus;
         }
+
+
         jsonStatus.setSuccess(true);
         jsonStatus.setMsg("保存成功!");
         return jsonStatus;
     }
 
     @Override
-    public JsonStatus saveWorkGroupRoles(long workGroupId, String roleId) {
-        JsonStatus jsonStatus=new JsonStatus();
-        try {
-            workGroupRoleBeanDao.deleteByWorkGroupId(workGroupId);
-            String userName=getShiroService().getCurrentUserName();
-            if (StringUtils.isNotEmpty(roleId)) {
-                if (roleId.indexOf(",") != -1) {
+    public JsonStatus saveWorkGroupRoles(List ids) {
+        JsonStatus jsonStatus = new JsonStatus();
+
+        if (ids == null || ids.size() != 2) {
+            jsonStatus.setFailure(true);
+            jsonStatus.setMsg("保存失败!");
+            return jsonStatus;
+        } else {
+            try {
+                long workGroupId = Long.valueOf(ids.get(0).toString());
+                String roleId = ids.get(1).toString();
+
+                workGroupRoleBeanDao.deleteByWorkGroupId(workGroupId);
+                String userName = getShiroService().getCurrentUserName();
+                if (StringUtils.isNotEmpty(roleId)) {
                     String[] roleIds = roleId.split(",");
                     for (String _roleId : roleIds) {
                         if (StringUtils.isNotEmpty(_roleId.trim())) {
@@ -180,30 +191,22 @@ public class WorkGroupBeanServiceImpl extends ShiroGenericBizServiceImpl<IWorkGr
                             workGroupRoleBeanDao.save(workGroupRoleBean);
                         }
                     }
-                }else{
-                    if (StringUtils.isNotEmpty(roleId.trim())) {
-                        WorkGroupRoleBean workGroupRoleBean = new WorkGroupRoleBean();
-                        workGroupRoleBean.setCreateBy(userName);
-                        workGroupRoleBean.setUpdateBy(userName);
-                        workGroupRoleBean.setGroupId(workGroupId);
-                        workGroupRoleBean.setRoleId(Long.parseLong(roleId));
-                        workGroupRoleBeanDao.save(workGroupRoleBean);
-                    }
                 }
+            } catch (Exception e) {
+                e.printStackTrace();
+                jsonStatus.setFailure(true);
+                jsonStatus.setMsg("保存失败!");
+                return jsonStatus;
             }
-        }catch(Exception e){
-            e.printStackTrace();
-            jsonStatus.setFailure(true);
-            jsonStatus.setMsg("保存失败!");
+
+            jsonStatus.setSuccess(true);
+            jsonStatus.setMsg("保存成功!");
             return jsonStatus;
         }
-        jsonStatus.setSuccess(true);
-        jsonStatus.setMsg("保存成功!");
-        return jsonStatus;
     }
 
-    @Override
-    public List<WorkGroupUserBean> getWorkGroupUserBeanByUserId(long userId) {
-        return workGroupUserBeanDao.find("select wgu from WorkGroupUserBean wgu where wgu.userId=?1",userId);
+        @Override
+        public List<WorkGroupUserBean> getWorkGroupUserBeanByUserId ( long userId){
+            return workGroupUserBeanDao.find("select wgu from WorkGroupUserBean wgu where wgu.userId=?1", userId);
+        }
     }
-}
